Fix obsidan instruments

+++ src/main/java/net/rezolv/obsidanum/item/custom/ObsidanAxe.java
@@ -23,6 +23,7 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
 
 
 public class ObsidanAxe extends AxeItem {
@@ -32,7 +33,7 @@ public class ObsidanAxe extends AxeItem {
 
     private boolean activated = false;
     private long lastActivationTime = 0;
-    private static final long COOLDOWN_DURATION = 120 * 20; // 60 seconds in ticks
+    private static final long COOLDOWN_DURATION = 1 * 20; // 60 seconds in ticks
     private static final long ACTIVATION_DURATION = 5 * 20; // 5 seconds in ticks
 
 
@@ -112,23 +113,25 @@ public class ObsidanAxe extends AxeItem {
         return super.mineBlock(stack, world, state, pos, entity);
     }
 
-    private void chainBreak(Level world, BlockPos pos, Player player, ItemStack stack) {
-        Set<BlockPos> visited = new HashSet<>();
-        breakBlock(world, pos, player, stack, visited);
-    }
-
-    private void breakBlock(Level world, BlockPos pos, Player player, ItemStack stack, Set<BlockPos> visited) {
-        if (visited.contains(pos)) return;
+    private void breakBlock(Level world, BlockPos pos, Player player, ItemStack stack, Set<BlockPos> visited, AtomicInteger blockBreakCount) {
+        if (visited.contains(pos) || blockBreakCount.get() >= 150) return;
         visited.add(pos);
 
         BlockState state = world.getBlockState(pos);
         Block block = state.getBlock();
 
-        if (block.defaultBlockState().is(MINEABLE_LOGS_TAG) || block.defaultBlockState().is(MINEABLE_LEAVES_TAG)) {
+        if (block.defaultBlockState().is(MINEABLE_LOGS_TAG) || block.defaultBlockState().is(MINEABLE_LEAVES_TAG) || block == Blocks.NETHER_WART_BLOCK || block == Blocks.WARPED_WART_BLOCK || block == Blocks.SHROOMLIGHT) {
             world.destroyBlock(pos, true);
+            blockBreakCount.incrementAndGet(); // Увеличиваем счётчик разрушенных блоков
             for (BlockPos offset : BlockPos.betweenClosed(pos.offset(-1, -1, -1), pos.offset(1, 1, 1))) {
-                breakBlock(world, offset, player, stack, visited);
+                breakBlock(world, offset, player, stack, visited, blockBreakCount);
             }
         }
+    }
+
+    private void chainBreak(Level world, BlockPos pos, Player player, ItemStack stack) {
+        Set<BlockPos> visited = new HashSet<>();
+        AtomicInteger blockBreakCount = new AtomicInteger(0);
+        breakBlock(world, pos, player, stack, visited, blockBreakCount);
     }
 }
